requirements.txt
Flask==1.1.1
Flask-Login==0.5.0
flask-oidc==1.4.0
hvac==0.10.4
json5==0.9.1
jsonschema==3.2.0
mysql-connector-python==8.0.20
oauth2client==4.1.3
poly1305-aes==0.13.39
pyOpenSSL==19.1.0
PyYAML==5.3
pyspx
qiskit==0.23.1
requests==2.23.0
requests-ntlm==1.1.0
urllib3==1.25.8
